+++ page/detail/detail.module.js
@@ -16,6 +16,17 @@ export default function detailInit() {
         document.querySelector("#similar_apartments").innerHTML += components.apartment_card.func({id: apartment.id, title: apartment.name, image: apartment.thumbnail_image, description: apartment.description, price: apartment.price_per_night});
     };
 
+    // Set similar apartments not found
+    const setNoSimilarApartmentsFound = () => {
+        document.querySelector("#similar_apartments").innerHTML = `
+            <h2 class="text-xl text-black dark:text-white">No similar apartments found.</h2>
+        `;
+    };
+
+    const clearSimilarApartments = () => {
+        document.querySelector("#similar_apartments").innerHTML = "";
+    };
+
     // Get apartment by apartment parameter
     const params = new URLSearchParams(window.location.search);
     if (params.has("apartment")) {
@@ -34,51 +45,26 @@ export default function detailInit() {
                 const titleElement = document.querySelector("title");
                 if (titleElement) titleElement.innerText = apartment.name + " | wp-bnb";
 
+                // Clear similar apartment list
+                clearSimilarApartments();
+
                 // Load apartment view
                 renderApartmentDetailView(apartment);
 
                 // Load similar apartments into the list
+                let similarApartmentFound = false;
                 Object.values(apartment_list).forEach(similar_apartment => {
-                    if (apartment.max_guests <= similar_apartment.max_guests) renderApartmentCard(similar_apartment);
+                    if (apartment.max_guests <= similar_apartment.max_guests && apartment.id != similar_apartment.id) {
+                        similarApartmentFound = true;
+                        renderApartmentCard(similar_apartment);
+                    }
                 });
+
+                if (!similarApartmentFound) {
+                    setNoSimilarApartmentsFound();
+                }
             }
         }
     }
-
-    // let cart = [];
-    //
-    // // Hinzufügen eines Produkts zum Warenkorb
-    // $('body').on('click', '.add-to-cart', function() {
-    //     const productName = $(this).data('product');
-    //     const productPrice = parseFloat($(this).data('price'));
-    //     cart.push({ name: productName, price: productPrice });
-    //     updateCart();
-    // });
-    //
-    // // Warenkorb aktualisieren
-    // function updateCart() {
-    //     const cartItems = $('#cart-items');
-    //     cartItems.empty();
-    //     let total = 0;
-    //
-    //     cart.forEach(function(item, index) {
-    //         total += item.price;
-    //         cartItems.append(
-    //             `<li>
-    //                 ${item.name} - ${item.price.toFixed(2)}€
-    //                 <button class="remove" data-index="${index}">Entfernen</button>
-    //             </li>`
-    //         );
-    //     });
-    //
-    //     $('#total').text('Gesamtpreis: ' + total.toFixed(2) + '€');
-    // }
-    //
-    // // Produkt aus dem Warenkorb entfernen
-    // $(document).on('click', '.remove', function() {
-    //     const index = $(this).data('index');
-    //     cart.splice(index, 1);  // Produkt aus dem Array entfernen
-    //     updateCart();
-    // });
 };